Update base library to latest

+++ nativeui/win/text_edit_win.cc
@@ -156,7 +156,8 @@ RectF TextEdit::GetTextBounds() const {
   scoped_refptr<AttributedText> attributed_text =
       new AttributedText(GetWindowString(edit->hwnd()), TextFormat());
   attributed_text->SetFont(edit->font());
-  return attributed_text->GetBoundsFor(SizeF(GetBounds().width(), INT_MAX));
+  return attributed_text->GetBoundsFor(
+      SizeF(GetBounds().width(), static_cast<float>(INT_MAX)));
 }
 
 }  // namespace nu
